pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "dbcache"
version = "0.2.5"
description = "Incremental cache (CSV/Parquet) for SELECT with temporal partitioning and deduplication."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Seu Nome", email = "seu.email@example.com"}]
keywords = ["cache", "csv", "parquet", "database", "postgres", "etl", "timeseries"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "pandas>=2.0.0",
  "pyarrow>=13.0.0",
  "psycopg[binary]>=3.1.0; python_version>='3.8'",
  "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
postgres = ["psycopg>=3.1.0; python_version>='3.8'", "psycopg2-binary>=2.9.0"]
parquet = ["pyarrow>=13.0.0"]
dbcache = ["pyarrow>=13.0.0", "psycopg>=3.1.0; python_version>='3.8'", "psycopg2-binary>=2.9.0"]
all = ["pyarrow>=13.0.0", "psycopg>=3.1.0; python_version>='3.8'", "psycopg2-binary>=2.9.0"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
